+++ DOCS/CHANGELOG.md
@@ -27,8 +27,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 - structs:
   - namespaces: `Arch`, `ByteSearch`, `Char`, `Env`, `Mem`, `Ptr`, `Str`.
   - `CompressionMode`, `EncodingMode`, `Pnm`.
-  - standalone error types: `InvalidErrorConversion`, `ErrorNotImplemented`, `ErrorNotSupported`, `ErrorElementNotFound`, `ErrorInvalidAxisLength`, `ErrorKeyAlreadyExists`, `ErrorMismatchedDimensions`, `ErrorMismatchedIndices`, `ErrorMismatchedLength`, `ErrorNodeEmpty`, `ErrorNodeLinkNotSet`, `ErrorNodeLinkNotUnique`, `ErrorNotEnoughElements`, `ErrorNotEnoughSpace`, `ErrorOutOfBounds`, `ErrorOverflow`, `ErrorPartiallyAdded`.
-  - composite error types: `ErrorPartialSpace`.
+  - standalone error types: `InvalidErrorConversion`, `NotImplemented`, `NotSupported`, `ElementNotFound`, `InvalidAxisLength`, `KeyAlreadyExists`, `MismatchedDimensions`, `MismatchedIndices`, `MismatchedLength`, `NodeEmpty`, `NodeLinkNotSet`, `NodeLinkNotUnique`, `NotEnoughElements`, `NotEnoughSpace`, `OutOfBounds`, `Overflow`, `PartiallyAdded`.
+  - composite error types: `PartialSpace`.
   - `False`, `True`, `UnitBi`, `UnitSi`.
   - `HasherPengy`.
   - `Lgc16`.